# docs/openapi_fe.yaml
openapi: 3.0.3
info:
  title: RinnoGen API — Frontend-Facing
  description: |
    Public-facing endpoints consumed by the RinnoGen frontend.
    Admin, webhook, and internal service routes are excluded.

    **Response Envelope**:
    - Success: `{status, code, message, data}`
    - Error:   `{status, code, message, error, details}`

    **Auth**: Protected routes require `Authorization: Bearer <JWT>`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiSuccess:
      type: object
      required: [status, code, message, data]
      properties:
        status: { type: string, enum: [success] }
        code: { type: integer }
        message: { type: string }
        data: { type: object }
      example:
        status: success
        code: 200
        message: "Endpoint specific success message"
        data: {}
    ApiError:
      type: object
      required: [status, code, message, error]
      properties:
        status: { type: string, enum: [error] }
        code: { type: integer }
        message: { type: string }
        error: { type: string }
        details: { type: object, nullable: true }
      example:
        status: error
        code: 400
        message: Invalid request body
        error: BAD_REQUEST
        details: null

    # --- Auth ---
...
    Subject:
      type: object
      properties:
        id: { type: integer, example: 1 }
        title: { type: string, example: JavaScript }
        slug: { type: string, example: javascript }
        color: { type: string, nullable: true, example: "#f7df1e" }
        languageId: { type: integer, nullable: true, example: 1 }
    SubjectWithDetail:
      type: object
      properties:
        subject: { $ref: "#/components/schemas/Subject" }
    SubjectListData:
      type: object
      properties:
        subjects:
          type: array
          items: { $ref: "#/components/schemas/Subject" }
    Session:
      type: object
      properties:
        id: { type: integer, example: 1 }
        subjectId: { type: integer, example: 1 }
        title: { type: string, example: "Session 01 - Basics" }
        description:
          {
            type: string,
            nullable: true,
            example: "Learn the fundamentals of JavaScript",
          }
        orderIndex: { type: integer, example: 1 }
    SessionListData:
      type: object
      properties:
        sessions:
          type: array
          items: { $ref: "#/components/schemas/Session" }
    Lesson:
      type: object
      properties:
        id: { type: integer, example: 1 }
        subjectSessionId: { type: integer, example: 1 }
        title: { type: string, example: "Lesson 01 - Input and Output" }
        contentMd:
          {
            type: string,
            nullable: true,
            example: "# Input and Output\n\nLearn how to read input...",
          }
        orderIndex: { type: integer, example: 1 }
    LessonDetail:
      type: object
      properties:
        lesson: { $ref: "#/components/schemas/Lesson" }
    LessonListData:
      type: object
      properties:
        lessons:
          type: array
          items: { $ref: "#/components/schemas/Lesson" }
    ProblemListItem:
      type: object
      properties:
        id: { type: integer, example: 1 }
        slug: { type: string, example: sum-two-numbers }
        title: { type: string, example: "Sum Two Numbers" }
        difficulty: { type: string, enum: [Easy, Medium, Hard], example: Easy }
        orderIndex: { type: integer, example: 1 }
        acceptanceRate: { type: number, example: 75.5 }
    ProblemListData:
      type: object
      properties:
        problems:
          type: array
          items: { $ref: "#/components/schemas/ProblemListItem" }

    # --- Problems ---
    ProblemDetail:
      type: object
      properties:
        executionMode:
          type: string
          description: The execution mode (e.g. stdin_stdout or function)
        functionName:
          type: string
          nullable: true
          description: Name of the function to be implemented
        initialCode:
          type: string
          nullable: true
          description: Starting code template for the user
        id: { type: integer, example: 1 }
        slug: { type: string, example: sum-two-numbers }
        title: { type: string, example: "Sum Two Numbers" }
        difficulty: { type: string, example: Easy }
        problemMd:
          {
            type: string,
            example: "## Sum Two Numbers\n\nGiven two integers, return their sum.",
          }
        timeLimitMs: { type: integer, example: 1000 }
        memoryLimitMb: { type: integer, example: 128 }
        acceptanceRate: { type: number, example: 75.5 }
        sampleTestCases:
          type: array
          items:
            type: object
            properties:
              input: { type: string, example: "2 3" }
              output: { type: string, example: "5" }
    ProblemDetailData:
      type: object
      properties:
        problem: { $ref: "#/components/schemas/ProblemDetail" }
    TestCase:
      type: object
      properties:
        id: { type: integer, example: 1 }
        problemId: { type: integer, example: 1 }
        inputData: { type: string, nullable: true, example: "2 3" }
        expectedOutput: { type: string, example: "5" }
        orderIndex: { type: integer, example: 1 }
    TestCaseListData:
      type: object
      properties:
        testCases:
          type: array
          items: { $ref: "#/components/schemas/TestCase" }

    # --- Runner ---
...
  responses:
    BadRequest:
      description: Invalid request (malformed body or missing required fields)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example:
            status: error
            code: 400
            message: Invalid request body
            error: BAD_REQUEST
            details: null
...
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example:
            status: error
            code: 404
            message: Not found
            error: NOT_FOUND
            details: null
...
    InternalError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example:
            status: error
            code: 500
            message: Internal server error
            error: INTERNAL_ERROR
            details: null
paths:
...
  /subjects:
    get:
      tags: [Curriculum]
      security: []
      summary: List published subjects
      description: Returns all published subjects ordered by ID. No authentication required.
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/SubjectListData" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  subjects:
                    - id: 1
                      title: JavaScript
                      slug: javascript
                      color: "#f7df1e"
                      languageId: 1
                    - id: 2
                      title: Python
                      slug: python
                      color: "#3572A5"
                      languageId: 2
        "500": { $ref: "#/components/responses/InternalError" }
  /subjects/{slug}:
    get:
      tags: [Curriculum]
      security: []
      summary: Get subject by slug
      description: Returns a single published subject identified by its URL-safe slug.
      parameters:
        - {
            name: slug,
            in: path,
            required: true,
            schema: { type: string },
            example: javascript,
          }
      responses:
        "200":
          description: Subject details
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/SubjectWithDetail" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  subject:
                    id: 1
                    title: JavaScript
                    slug: javascript
                    color: "#f7df1e"
                    languageId: 1
        "404": { $ref: "#/components/responses/NotFound" }
  /subjects/{slug}/sessions:
    get:
      tags: [Curriculum]
      security: []
      summary: List sessions for a subject
      description: Returns all sessions belonging to the subject, ordered by orderIndex ascending.
      parameters:
        - {
            name: slug,
            in: path,
            required: true,
            schema: { type: string },
            example: javascript,
          }
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/SessionListData" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  sessions:
                    - id: 1
                      subjectId: 1
                      title: "Session 01 - Basics"
                      description: "Learn the fundamentals of JavaScript"
                      orderIndex: 1
                    - id: 2
                      subjectId: 1
                      title: "Session 02 - Functions"
                      description: "Deep dive into functions"
                      orderIndex: 2
        "404": { $ref: "#/components/responses/NotFound" }

  # ──────────────────────────────────────────────
  # Curriculum — Lessons
  # ──────────────────────────────────────────────
  /sessions/{id}/lessons:
    get:
      tags: [Curriculum]
      security: []
      summary: List lessons for a session
      description: Returns all lessons in a session, ordered by orderIndex ascending.
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: integer },
            example: 1,
          }
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/LessonListData" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  lessons:
                    - id: 1
                      subjectSessionId: 1
                      title: "Lesson 01 - Input and Output"
                      contentMd: "# Input and Output\n\nLearn how to read input and print output."
                      orderIndex: 1
                    - id: 2
                      subjectSessionId: 1
                      title: "Lesson 02 - Variables"
                      contentMd: "# Variables\n\nLearn about variable declarations."
                      orderIndex: 2
        "400": { $ref: "#/components/responses/BadRequest" }
  /lessons/{id}:
    get:
      tags: [Curriculum]
      security: []
      summary: Get lesson by ID
      description: Returns a single lesson including its Markdown content.
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: integer },
            example: 1,
          }
      responses:
        "200":
          description: Lesson details
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/LessonDetail" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  lesson:
                    id: 1
                    subjectSessionId: 1
                    title: "Lesson 01 - Input and Output"
                    contentMd: "# Input and Output\n\nLearn how to read input and print output."
                    orderIndex: 1
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /lessons/{id}/problems:
    get:
      tags: [Curriculum]
      security: []
      summary: List problems for a lesson
      description: Returns published problems associated with a lesson, ordered by the lesson-specific orderIndex.
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: integer },
            example: 1,
          }
      responses:
        "200":
          description: List of problems
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/ProblemListData" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  problems:
                    - id: 1
                      slug: sum-two-numbers
                      title: "Sum Two Numbers"
                      difficulty: Easy
                      orderIndex: 1
                      acceptanceRate: 75.5
                    - id: 2
                      slug: hello-world
                      title: "Hello World"
                      difficulty: Easy
                      orderIndex: 2
                      acceptanceRate: 90.0
        "400": { $ref: "#/components/responses/BadRequest" }

  # ──────────────────────────────────────────────
  /problems/daily-challenge:
    get:
      tags: [Problems]
      security: []
      summary: Get daily challenge
      description: Returns today's daily challenge problem. If one hasn't been selected yet today, selects an unused published problem.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          date:
                            type: string
                            format: date
                          problem:
                            $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No daily challenge available
  /problems/{slug}:
    get:
      tags: [Problems]
      security: []
      summary: Get problem by slug
      description: Returns full problem details including Markdown description, time/memory limits, acceptance rate, and sample test cases.
      parameters:
        - {
            name: slug,
            in: path,
            required: true,
            schema: { type: string },
            example: sum-two-numbers,
          }
      responses:
        "200":
          description: Problem details
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/ProblemDetailData" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  problem:
                    id: 1
                    slug: sum-two-numbers
                    title: "Sum Two Numbers"
                    difficulty: Easy
                    problemMd: "## Sum Two Numbers\n\nGiven two integers, return their sum."
                    timeLimitMs: 1000
                    memoryLimitMb: 128
                    acceptanceRate: 75.5
                    sampleTestCases:
                      - input: "2 3"
                        output: "5"
                      - input: "-1 1"
                        output: "0"
        "404": { $ref: "#/components/responses/NotFound" }
  /problems/{id}/test-cases:
    get:
      tags: [Problems]
      security: []
      summary: List sample test cases for a problem
      description: Returns only sample (visible) test cases for a problem. Hidden test cases are never exposed to clients.
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: integer },
            example: 1,
          }
        - {
            name: visibility,
            in: query,
            schema: { type: string, enum: [sample], default: sample },
          }
      responses:
        "200":
          description: List of sample test cases
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/TestCaseListData" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  testCases:
                    - id: 1
                      problemId: 1
                      inputData: "2 3"
                      expectedOutput: "5"
                      orderIndex: 1
                    - id: 2
                      problemId: 1
                      inputData: "-1 1"
                      expectedOutput: "0"
                      orderIndex: 2
        "400": { $ref: "#/components/responses/BadRequest" }

  # ──────────────────────────────────────────────
  # Runner
  # ──────────────────────────────────────────────
...